refactor: move data to the separate file
