+++ src/components/shared/request-form/request-form.jsx
@@ -16,249 +16,7 @@ import CheckIcon from 'icons/check.inline.svg';
 import ChevronIcon from 'icons/chevron-down.inline.svg';
 import { emailRegexp } from 'utils/forms';
 
-const REGIONS = {
-  title: 'Request a new Provider / Region',
-  description:
-    "Looking for Neon in a different cloud provider or region? Select your preference below, and we'll notify you as soon as it's available.",
-  placeholder: 'Select a region',
-  buttonText: 'Request',
-  options: [
-    { id: 'af-south-1', name: 'AWS - Africa (Cape Town)' },
-    { id: 'ap-east-1', name: 'AWS - Asia Pacific (Hong Kong)' },
-    { id: 'ap-northeast-1', name: 'AWS - Asia Pacific (Tokyo)' },
-    { id: 'ap-northeast-2', name: 'AWS - Asia Pacific (Seoul)' },
-    { id: 'ap-northeast-3', name: 'AWS - Asia Pacific (Osaka)' },
-    { id: 'ap-south-1', name: 'AWS - Asia Pacific (Mumbai)' },
-    { id: 'ap-south-2', name: 'AWS - Asia Pacific (Hyderabad)' },
-    { id: 'ap-southeast-3', name: 'AWS - Asia Pacific (Jakarta)' },
-    { id: 'ap-southeast-4', name: 'AWS - Asia Pacific (Melbourne)' },
-    { id: 'ca-central-1', name: 'AWS - Canada (Central)' },
-    { id: 'ca-west-1', name: 'AWS - Canada West (Calgary)' },
-    { id: 'cn-north-1', name: 'AWS - China (Beijing)' },
-    { id: 'cn-northwest-1', name: 'AWS - China (Ningxia)' },
-    { id: 'eu-central-2', name: 'AWS - Europe (Zurich)' },
-    { id: 'eu-north-1', name: 'AWS - Europe (Stockholm)' },
-    { id: 'eu-south-1', name: 'AWS - Europe (Milan)' },
-    { id: 'eu-south-2', name: 'AWS - Europe (Spain)' },
-    { id: 'eu-west-1', name: 'AWS - Europe (Ireland)' },
-    { id: 'eu-west-3', name: 'AWS - Europe (Paris)' },
-    { id: 'il-central-1', name: 'AWS - Israel (Tel Aviv)' },
-    { id: 'me-central-1', name: 'AWS - Middle East (UAE)' },
-    { id: 'me-south-1', name: 'AWS - Middle East (Bahrain)' },
-    { id: 'us-gov-east-1', name: 'AWS - AWS GovCloud (US-East)' },
-    { id: 'us-gov-west-1', name: 'AWS - AWS GovCloud (US-West)' },
-    { id: 'us-west-1', name: 'AWS - US West (N. California)' },
-    { id: 'eastus', name: 'Azure - (US) East US' },
-    { id: 'southcentralus', name: 'Azure - (US) South Central US' },
-    { id: 'westus2', name: 'Azure - (US) West US 2' },
-    { id: 'australiaeast', name: 'Azure - (Asia Pacific) Australia East' },
-    { id: 'southeastasia', name: 'Azure - (Asia Pacific) Southeast Asia' },
-    { id: 'northeurope', name: 'Azure - (Europe) North Europe' },
-    { id: 'swedencentral', name: 'Azure - (Europe) Sweden Central' },
-    { id: 'uksouth', name: 'Azure - (Europe) UK South' },
-    { id: 'westeurope', name: 'Azure - (Europe) West Europe' },
-    { id: 'centralus', name: 'Azure - (US) Central US' },
-    { id: 'southafricanorth', name: 'Azure - (Africa) South Africa North' },
-    { id: 'centralindia', name: 'Azure - (Asia Pacific) Central India' },
-    { id: 'eastasia', name: 'Azure - (Asia Pacific) East Asia' },
-    { id: 'japaneast', name: 'Azure - (Asia Pacific) Japan East' },
-    { id: 'koreacentral', name: 'Azure - (Asia Pacific) Korea Central' },
-    { id: 'canadacentral', name: 'Azure - (Canada) Canada Central' },
-    { id: 'francecentral', name: 'Azure - (Europe) France Central' },
-    { id: 'italynorth', name: 'Azure - (Europe) Italy North' },
-    { id: 'norwayeast', name: 'Azure - (Europe) Norway East' },
-    { id: 'polandcentral', name: 'Azure - (Europe) Poland Central' },
-    { id: 'spaincentral', name: 'Azure - (Europe) Spain Central' },
-    { id: 'switzerlandnorth', name: 'Azure - (Europe) Switzerland North' },
-    { id: 'mexicocentral', name: 'Azure - (Mexico) Mexico Central' },
-    { id: 'uaenorth', name: 'Azure - (Middle East) UAE North' },
-    { id: 'brazilsouth', name: 'Azure - (South America) Brazil South' },
-    { id: 'israelcentral', name: 'Azure - (Middle East) Israel Central' },
-    { id: 'qatarcentral', name: 'Azure - (Middle East) Qatar Central' },
-    { id: 'centralusstage', name: 'Azure - (US) Central US (Stage)' },
-    { id: 'eastusstage', name: 'Azure - (US) East US (Stage)' },
-    { id: 'eastus2stage', name: 'Azure - (US) East US 2 (Stage)' },
-    { id: 'northcentralusstage', name: 'Azure - (US) North Central US (Stage)' },
-    { id: 'southcentralusstage', name: 'Azure - (US) South Central US (Stage)' },
-    { id: 'westusstage', name: 'Azure - (US) West US (Stage)' },
-    { id: 'westus2stage', name: 'Azure - (US) West US 2 (Stage)' },
-    { id: 'asia', name: 'Azure - Asia' },
-    { id: 'asiapacific', name: 'Azure - Asia Pacific' },
-    { id: 'australia', name: 'Azure - Australia' },
-    { id: 'brazil', name: 'Azure - Brazil' },
-    { id: 'canada', name: 'Azure - Canada' },
-    { id: 'europe', name: 'Azure - Europe' },
-    { id: 'france', name: 'Azure - France' },
-    { id: 'germany', name: 'Azure - Germany' },
-    { id: 'global', name: 'Azure - Global' },
-    { id: 'india', name: 'Azure - India' },
-    { id: 'israel', name: 'Azure - Israel' },
-    { id: 'italy', name: 'Azure - Italy' },
-    { id: 'japan', name: 'Azure - Japan' },
-    { id: 'korea', name: 'Azure - Korea' },
-    { id: 'newzealand', name: 'Azure - New Zealand' },
-    { id: 'norway', name: 'Azure - Norway' },
-    { id: 'poland', name: 'Azure - Poland' },
-    { id: 'qatar', name: 'Azure - Qatar' },
-    { id: 'singapore', name: 'Azure - Singapore' },
-    { id: 'southafrica', name: 'Azure - South Africa' },
-    { id: 'sweden', name: 'Azure - Sweden' },
-    { id: 'switzerland', name: 'Azure - Switzerland' },
-    { id: 'uae', name: 'Azure - United Arab Emirates' },
-    { id: 'uk', name: 'Azure - United Kingdom' },
-    { id: 'unitedstates', name: 'Azure - United States' },
-    { id: 'unitedstateseuap', name: 'Azure - United States EUAP' },
-    { id: 'eastasiastage', name: 'Azure - (Asia Pacific) East Asia (Stage)' },
-    { id: 'southeastasiastage', name: 'Azure - (Asia Pacific) Southeast Asia (Stage)' },
-    { id: 'brazilus', name: 'Azure - (South America) Brazil US' },
-    { id: 'eastusstg', name: 'Azure - (US) East US STG' },
-    { id: 'northcentralus', name: 'Azure - (US) North Central US' },
-    { id: 'westus', name: 'Azure - (US) West US' },
-    { id: 'japanwest', name: 'Azure - (Asia Pacific) Japan West' },
-    { id: 'jioindiawest', name: 'Azure - (Asia Pacific) Jio India West' },
-    { id: 'centraluseuap', name: 'Azure - (US) Central US EUAP' },
-    { id: 'eastus2euap', name: 'Azure - (US) East US 2 EUAP' },
-    { id: 'westcentralus', name: 'Azure - (US) West Central US' },
-    { id: 'southafricawest', name: 'Azure - (Africa) South Africa West' },
-    { id: 'australiacentral', name: 'Azure - (Asia Pacific) Australia Central' },
-    { id: 'australiacentral2', name: 'Azure - (Asia Pacific) Australia Central 2' },
-    { id: 'australiasoutheast', name: 'Azure - (Asia Pacific) Australia Southeast' },
-    { id: 'jioindiacentral', name: 'Azure - (Asia Pacific) Jio India Central' },
-    { id: 'koreasouth', name: 'Azure - (Asia Pacific) Korea South' },
-    { id: 'southindia', name: 'Azure - (Asia Pacific) South India' },
-    { id: 'westindia', name: 'Azure - (Asia Pacific) West India' },
-    { id: 'canadaeast', name: 'Azure - (Canada) Canada East' },
-    { id: 'francesouth', name: 'Azure - (Europe) France South' },
-    { id: 'germanynorth', name: 'Azure - (Europe) Germany North' },
-    { id: 'norwaywest', name: 'Azure - (Europe) Norway West' },
-    { id: 'switzerlandwest', name: 'Azure - (Europe) Switzerland West' },
-    { id: 'ukwest', name: 'Azure - (Europe) UK West' },
-    { id: 'uaecentral', name: 'Azure - (Middle East) UAE Central' },
-    { id: 'brazilsoutheast', name: 'Azure - (South America) Brazil Southeast' },
-    { id: 'asia-south1', name: 'GCP - Mumbai' },
-    { id: 'asia-south2', name: 'GCP - Delhi' },
-    { id: 'asia-east1', name: 'GCP - Changhua County' },
-    { id: 'asia-east2', name: 'GCP - Hong Kong' },
-    { id: 'asia-northeast1', name: 'GCP - Tokyo' },
-    { id: 'asia-northeast2', name: 'GCP - Osaka' },
-    { id: 'asia-northeast3', name: 'GCP - Seoul' },
-    { id: 'asia-southeast1', name: 'GCP - Jurong West' },
-    { id: 'australia-southeast1', name: 'GCP - Sydney' },
-    { id: 'australia-southeast2', name: 'GCP - Melbourne' },
-    { id: 'europe-central2', name: 'GCP - Warsaw' },
-    { id: 'europe-north2', name: 'GCP - Hamina' },
-    { id: 'europe-southwest1', name: 'GCP - Madrid' },
-    { id: 'europe-west1', name: 'GCP - St. Ghislain' },
-    { id: 'europe-west2', name: 'GCP - London' },
-    { id: 'europe-west3', name: 'GCP - Frankfurt' },
-    { id: 'europe-west4', name: 'GCP - Eemshaven' },
-    { id: 'europe-west6', name: 'GCP - Zurich' },
-    { id: 'europe-west8', name: 'GCP - Milan' },
-    { id: 'europe-west9', name: 'GCP - Paris' },
-    { id: 'northamerica-northeast1', name: 'GCP - Montréal' },
-    { id: 'northamerica-northeast2', name: 'GCP - Toronto' },
-    { id: 'southamerica-east1', name: 'GCP - Osasco' },
-    { id: 'us-central1', name: 'GCP - Council Bluffs' },
-    { id: 'us-east1', name: 'GCP - Moncks Corner' },
-    { id: 'us-east4', name: 'GCP - Ashburn' },
-    { id: 'us-west-1', name: 'GCP - Oregon' },
-    { id: 'us-west2', name: 'GCP - Los Angeles' },
-    { id: 'us-west3', name: 'GCP - Salt Lake City' },
-  ],
-  eventName: 'Region Requested',
-  eventProps: {
-    name: 'region_name',
-    id: 'region_id',
-  },
-};
-
-const EXTENSIONS = {
-  title: 'Request a new extension',
-  description: 'Looking for a specific extension in Neon? Suggest one using this form.',
-  placeholder: 'Select an extension',
-  buttonText: 'Request',
-  options: [
-    { name: 'aiven_extras' },
-    { name: 'adminpack' },
-    { name: 'amcheck' },
-    { name: 'auth_delay' },
-    { name: 'auto_explain' },
-    { name: 'aws_commons' },
-    { name: 'aws_lambda' },
-    { name: 'aws_s3' },
-    { name: 'basebackup_to_shell' },
-    { name: 'basic_archive' },
-    { name: 'bool_plperl' },
-    { name: 'bool_plperlu' },
-    { name: 'cstore_fdw' },
-    { name: 'Citus' },
-    { name: 'dblink' },
-    { name: 'dict_xsyn' },
-    { name: 'file_fdw' },
-    { name: 'flow_control' },
-    { name: 'hstore_plperl' },
-    { name: 'hstore_plperlu' },
-    { name: 'http' },
-    { name: 'jsonb_plperl' },
-    { name: 'jsonb_plperlu' },
-    { name: 'log_fdw' },
-    { name: 'mongo_fdw' },
-    { name: 'Multicorn' },
-    { name: 'mysql_fdw' },
-    { name: 'old_snapshot' },
-    { name: 'oracle_fdw' },
-    { name: 'orafce' },
-    { name: 'pageinspect' },
-    { name: 'passwordcheck' },
-    { name: 'pg_anon' },
-    { name: 'pg_bigm' },
-    { name: 'pg_buffercache' },
-    { name: 'pg_freespacemap' },
-    { name: 'pg_parquet' },
-    { name: 'pg_proctab' },
-    { name: 'pg_similarity' },
-    { name: 'pgSphere' },
-    { name: 'pg_surgery' },
-    { name: 'pg_transport' },
-    { name: 'pg_visibility' },
-    { name: 'pg_walinspect' },
-    { name: 'pg_bulkload' },
-    { name: 'pg_prometheus' },
-    { name: 'pglogical' },
-    { name: 'pldebugger' },
-    { name: 'pljava' },
-    { name: 'plperl' },
-    { name: 'plperlu' },
-    { name: 'plprofiler' },
-    { name: 'pltcl' },
-    { name: 'PL/Proxy' },
-    { name: 'PostPic' },
-    { name: 'Postgres_fdw' },
-    { name: 'postgis_legacy' },
-    { name: 'pgaudit' },
-    { name: 'PgMemcache' },
-    { name: 'rds_tools' },
-    { name: 'sepgsql' },
-    { name: 'sslinfo' },
-    { name: 'tds_fdw' },
-    { name: 'test_decoding' },
-    { name: 'test_parser' },
-    { name: 'Texcaller' },
-    { name: 'timetravel' },
-    { name: 'uuid' },
-    { name: 'ZomboDB' },
-  ],
-  extendedOptions: true,
-  eventName: 'Extension Requested',
-  eventProps: {
-    name: 'extension_name',
-  },
-};
-
-const DATA = {
-  region: REGIONS,
-  extension: EXTENSIONS,
-};
+import DATA from './data';
 
 function getCookie(name) {
   if (typeof document === 'undefined') return null;
@@ -341,7 +99,6 @@ const RequestForm = ({ type }) => {
               value={selected}
               immediate
               onChange={(value) => {
-                console.log('Selected:', value);
                 setSelected(value);
               }}
               onClose={() => setQuery('')}
